+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### next
 - fix handling of rules starting with '/' in the global gitignore - Fix #321
+- alt-c now mapped to the new :copy_line verb which, when in tree, puts the selected path in the clipboard and, when in text preview, puts the selected text line in the clipboard - Fix #322
 
 <a name="v1.1.10"></a>
 ### v1.1.10 - 2020-12-24